Reorganize backend structure
