+++ backend/src/main/java/org/hackbrooklyn/plaza/controller/ApplyController.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.hackbrooklyn.plaza.exception.MultipartFormDataValidationException;
 import org.hackbrooklyn.plaza.model.SubmittedApplication;
 import org.hackbrooklyn.plaza.service.ApplyService;
 import org.springframework.beans.factory.annotation.Autowired;
