+++ backend/src/main/java/org/hackbrooklyn/plaza/controller/RestControllerExceptionHandler.java
@@ -1,7 +1,7 @@
 package org.hackbrooklyn.plaza.controller;
 
 import com.fasterxml.jackson.core.JsonParseException;
-import org.hackbrooklyn.plaza.service.*;
+import org.hackbrooklyn.plaza.exception.*;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.BadCredentialsException;
