+++ backend/src/main/java/org/hackbrooklyn/plaza/repository/SubmittedApplicationRepository.java
@@ -1,7 +1,7 @@
 package org.hackbrooklyn.plaza.repository;
 
 import org.hackbrooklyn.plaza.model.SubmittedApplication;
-import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository .JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
