--- backend/src/main/java/org/hackbrooklyn/plaza/service/AccountAlreadyActivatedException.java
@@ -1,19 +1,0 @@
-package org.hackbrooklyn.plaza.service;
-
-public class AccountAlreadyActivatedException extends RuntimeException {
-
-    public AccountAlreadyActivatedException() {
-    }
-
-    public AccountAlreadyActivatedException(String message) {
-        super(message);
-    }
-
-    public AccountAlreadyActivatedException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public AccountAlreadyActivatedException(Throwable cause) {
-        super(cause);
-    }
-}
