--- backend/src/main/java/org/hackbrooklyn/plaza/service/ApplicationNotFoundException.java
@@ -1,19 +1,0 @@
-package org.hackbrooklyn.plaza.service;
-
-public class ApplicationNotFoundException extends RuntimeException {
-
-    public ApplicationNotFoundException() {
-    }
-
-    public ApplicationNotFoundException(String message) {
-        super(message);
-    }
-
-    public ApplicationNotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public ApplicationNotFoundException(Throwable cause) {
-        super(cause);
-    }
-}
