--- backend/src/main/java/org/hackbrooklyn/plaza/service/InvalidKeyException.java
@@ -1,19 +1,0 @@
-package org.hackbrooklyn.plaza.service;
-
-public class InvalidKeyException extends RuntimeException {
-
-    public InvalidKeyException() {
-    }
-
-    public InvalidKeyException(String message) {
-        super(message);
-    }
-
-    public InvalidKeyException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public InvalidKeyException(Throwable cause) {
-        super(cause);
-    }
-}
