--- backend/src/main/java/org/hackbrooklyn/plaza/service/InvalidTokenException.java
@@ -1,19 +1,0 @@
-package org.hackbrooklyn.plaza.service;
-
-public class InvalidTokenException extends RuntimeException {
-
-    public InvalidTokenException() {
-    }
-
-    public InvalidTokenException(String message) {
-        super(message);
-    }
-
-    public InvalidTokenException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public InvalidTokenException(Throwable cause) {
-        super(cause);
-    }
-}
