--- backend/src/main/java/org/hackbrooklyn/plaza/service/PlazaUserDetailsService.java
@@ -1,26 +1,0 @@
-package org.hackbrooklyn.plaza.service;
-
-import org.hackbrooklyn.plaza.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-public class PlazaUserDetailsService implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    @Autowired
-    public PlazaUserDetailsService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return userRepository.findByEmail(username).orElseThrow(() ->
-                new UsernameNotFoundException("No user with the email \"" + username + "\" was found.")
-        );
-    }
-}
