--- backend/src/main/java/org/hackbrooklyn/plaza/service/RejectedFileTypeException.java
@@ -1,18 +1,0 @@
-package org.hackbrooklyn.plaza.service;
-
-public class RejectedFileTypeException extends RuntimeException {
-    public RejectedFileTypeException() {
-    }
-
-    public RejectedFileTypeException(String message) {
-        super(message);
-    }
-
-    public RejectedFileTypeException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public RejectedFileTypeException(Throwable cause) {
-        super(cause);
-    }
-}
