--- backend/src/main/java/org/hackbrooklyn/plaza/service/SendGridException.java
@@ -1,19 +1,0 @@
-package org.hackbrooklyn.plaza.service;
-
-public class SendGridException extends RuntimeException {
-
-    public SendGridException() {
-    }
-
-    public SendGridException(String message) {
-        super(message);
-    }
-
-    public SendGridException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public SendGridException(Throwable cause) {
-        super(cause);
-    }
-}
