--- backend/src/main/java/org/hackbrooklyn/plaza/service/TokenDTO.java
@@ -1,10 +1,0 @@
-package org.hackbrooklyn.plaza.service;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class TokenDTO {
-    private String token;
-}
