--- backend/src/main/java/org/hackbrooklyn/plaza/service/UserNotFoundException.java
@@ -1,19 +1,0 @@
-package org.hackbrooklyn.plaza.service;
-
-public class UserNotFoundException extends RuntimeException {
-
-    public UserNotFoundException() {
-    }
-
-    public UserNotFoundException(String message) {
-        super(message);
-    }
-
-    public UserNotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public UserNotFoundException(Throwable cause) {
-        super(cause);
-    }
-}
