+++ backend/src/main/java/org/hackbrooklyn/plaza/util/JwtUtils.java
@@ -6,7 +6,7 @@ import io.jsonwebtoken.security.Keys;
 import io.jsonwebtoken.security.SignatureException;
 import lombok.extern.slf4j.Slf4j;
 import org.hackbrooklyn.plaza.model.User;
-import org.hackbrooklyn.plaza.service.TokenDTO;
+import org.hackbrooklyn.plaza.dto.TokenDTO;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
